fast_event_api/requirements.txt
redis==4.1.0
fastapi==0.61.1
orjson==3.6.4
requests==2.25.1
uvicorn==0.12.2
uvloop==0.16.0
kafka-python=2.0.2
